added bush, logo and title music
