+++ Guacamole/Assets/Scripts/ClickSound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c58d96e11cbef43a99818388a277c2b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
